+++ wger/exercises/tests/test_exercise_images.py
@@ -221,6 +221,20 @@ class ExerciseImageFromJsonSimpleTests(SimpleTestCase):
         self.assertFalse(img.is_main)
 
 
+class ExerciseImageThumbnailsApiTestCase(WgerTestCase):
+    """
+    Tests the thumbnails action of the exercise image API endpoint
+    """
+
+    def test_thumbnails_non_numeric_pk(self):
+        """A non-numeric pk yields a 404 instead of raising a 500"""
+        response = self.client.get(
+            reverse('exerciseimage-thumbnails', kwargs={'pk': 'not-a-number'})
+        )
+
+        self.assertEqual(response.status_code, 404)
+
+
 # TODO: add POST and DELETE tests
 class ExerciseImagesApiTestCase(
     ActstreamUpdateMixin,