+++ wger/nutrition/api/serializers.py
@@ -13,10 +13,14 @@
 # You should have received a copy of the GNU Affero General Public License
 # along with Workout Manager.  If not, see <http://www.gnu.org/licenses/>.
 
+# Standard Library
+import logging
+
 # Django
 from django.conf import settings
 
 # Third Party
+from easy_thumbnails.exceptions import EasyThumbnailsError
 from easy_thumbnails.files import get_thumbnailer
 from rest_framework import serializers
 
@@ -30,6 +34,10 @@ from wger.nutrition.models import (
     MealItem,
     NutritionPlan,
 )
+from wger.utils.url import make_absolute_url
+
+
+logger = logging.getLogger(__name__)
 
 
 class IngredientWeightUnitSerializer(serializers.ModelSerializer):
@@ -195,8 +203,12 @@ class IngredientInfoSerializer(serializers.ModelSerializer):
         thumbnailer = get_thumbnailer(obj.image.image)
         for alias in aliases:
             opts = settings.THUMBNAIL_ALIASES[''][alias]
-            thumb = thumbnailer.get_thumbnail(opts)
-            result[alias] = request.build_absolute_uri(thumb.url)
+            try:
+                thumb = thumbnailer.get_thumbnail(opts)
+            except (EasyThumbnailsError, OSError, ValueError):
+                logger.warning('Could not generate thumbnails for image %s', obj.image.pk)
+                return None
+            result[alias] = make_absolute_url(thumb.url, request)
         return result